+++ doc/whats_new.rst
@@ -74,9 +74,8 @@ New features
      cumulative gain (NDCG).
      :issue:`7739` by :user:`David Gasquez <davidgasquez>`.
 
-   - Added :class:`gaussian_process.kernels.SelectDimensionKernel`, which 
-     allows for applying an input kernel on specific features of input
-     X.
+   - Added :class:`gaussian_process.kernels.SelectDimensionKernel` by which
+     a kernel can be applied on specific features.
      :issue:`6868` by :user:`Behzad Tabibian <btabibian>`.
 
 Enhancements